Logout using redux observable

+++ src/App.jsx
@@ -22,12 +22,27 @@ import AllHistorLogins from "./components/Pages/AllHistoryLogins/AllHistoryLogin
 import LoadingIndicatorBar from "./components/Widgets/LoadingIndicatorBar/LoadingIndicatorBar";
 
 import "./App.scss";
+import Modal from "@material-ui/core/Modal/Modal";
+import ModalNotice from "./components/Widgets/ModalNotice/ModalNotice";
 
 
 class App extends Component {
 
   state = {
-    initialized: false
+    initialized: false,
+    modal_is_open: false,
+    modal_notice_text: {
+      title: 'Error',
+      notice: ''
+    }
+  };
+
+  _modalClose = () => {
+    this.setState( { modal_is_open: false } );
+  };
+
+  _modalOpen = () => {
+    this.setState( { modal_is_open: true } );
   };
 
   componentDidMount(){
@@ -80,7 +95,7 @@ class App extends Component {
       dispatch( appActions.appRouterChange( { header_mobile_show : true, header_menu_mobile_show: true } ) );//default should always true, it will be overridden at page/component part if it should be false
     }
 
-    if ( nextProps.is_app_initialized && !nextProps.is_logged_in ) {
+    if ( nextProps.app.is_app_initialized && this.props.app.is_logged_in && !nextProps.app.is_logged_in ) {
       biqHelper.localStorage.clear();
       dispatch( appActions.appStatesReset() );
       dispatch( appActions.appRedirectToAgen() );
@@ -90,15 +105,56 @@ class App extends Component {
     return true;
   }
 
-  componentDidUpdate(prevProps) {
+  componentDidUpdate(prevProps, prevState, snapshot) {
     // window.scrollTo(0,0);
+    let {dispatch} = this.props;
     let body = $('html, body');
     body.stop().animate({scrollTop:0}, 500, 'swing');
-    if( this.props.should_redirect_to_agen ) window.location = biqConfig.agen.url_base + '/#/login/default';
+    if( this.props.app.should_redirect_to_agen ) window.location = biqConfig.agen.url_base + '/#/login/default';
+
+    if( this.props.app.is_logging_out ) {
+
+      if (
+        biqHelper.utils.httpResponseIsSuccess(
+          biqHelper.JSON.pathValueGet( this.props.app.logout_response, 'status' )
+        )
+      ) {
+        dispatch( appActions.appLoggedOut() );
+      }
+
+      else if (
+        biqHelper.utils.httpResponseIsError(
+          biqHelper.JSON.pathValueGet( this.props.app.logout_response, 'status' )
+        )
+      ) {
+        let response_status = biqHelper.JSON.pathValueGet( this.props.app.logout_response, 'response.response_code.status' );
+        let response_text = biqHelper.JSON.pathValueGet( this.props.app.logout_response, 'response.response_code.message' );
+        let notice = `${response_status}, ${ response_text }`;
+
+        notice = biqHelper.utils.isNullSome( response_status, response_text ) ? 'Error logout tidak diketahui' : notice;
+        this.setState( {
+          modal_notice_text: {
+            title: 'Gagal logout',
+            notice: notice
+          }
+        });
+
+        this._modalOpen();
+      }
+
+    }
+
   }
 
   render() {
-    if ( !this.state.initialized ) return <LoadingIndicatorBar className={"loading-indicator-main"} style={{ top: '5px' }} barStyle={{background: '#1c94fc'}}/>;
+    if ( !this.state.initialized ) return <LoadingIndicatorBar className={"loading-indicator-main"} style={{ top: 0 }} barStyle={{background: '#1c94fc'}}/>;
+
+    if ( this.props.app.is_logging_out || !this.props.app.is_logged_in ) return (
+      <>
+        <Header/>
+        <LoadingIndicatorBar className={"loading-indicator-main"} barStyle={{background: '#1c94fc'}}/>
+      </>
+    );
 
     return (
         <React.Fragment>
@@ -106,7 +162,7 @@ class App extends Component {
           <Header/>
 
           {
-            this.props.loading_indicator_show && <LoadingIndicatorBar className={"loading-indicator-main"} barStyle={{background: '#1c94fc'}}/>
+            this.props.app.loading_indicator_show && <LoadingIndicatorBar className={"loading-indicator-main"} barStyle={{background: '#1c94fc'}}/>
           }
 
           <Switch>
@@ -118,6 +174,17 @@ class App extends Component {
             <Redirect from="/" to="/dashboard" />
           </Switch>
 
+
+          <Modal
+            open={this.state.modal_is_open}
+            onClose={this._modalClose}>
+
+            <div className="modal-inner">
+              <ModalNotice modalClose={this._modalClose} title={this.state.modal_notice_text.title} notice={this.state.modal_notice_text.notice}/>
+            </div>
+
+          </Modal>
+
         </React.Fragment>
     )
   }
@@ -125,11 +192,7 @@ class App extends Component {
 
 const mapStateToProps = ( state ) => {
   return {
-    is_app_initialized: state.app.is_app_initialized,
-    should_redirect_to_agen: state.app.should_redirect_to_agen,
-    is_logged_in: state.app.is_logged_in,
-    window_size: state.app.window_size,
-    loading_indicator_show: state.app.loading_indicator_show
+    app: state.app
   }
 };
